LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audiobuffersourcenode-interface/sub-sample-buffer-stitching-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "buffer-stitching-1" 
PASS Executing "buffer-stitching-2" 
PASS Audit report 
PASS > [buffer-stitching-1] Subsample buffer stitching, same rates 
PASS   Test 1: context.sampleRate is equal to 44100. 
PASS   Stitched sine-wave buffers at sample rate 44100 equals [0,0.06264829635620117,0.12505052983760834,0.1869613528251648,0.24813774228096008,0.308339387178421,0.36732959747314453,0.4248766601085663,0.48075437545776367,0.53474360704422,0.5866317749023438,0.6362154483795166,0.6832994818687439,0.7276992797851562,0.769240140914917,0.8077590465545654...] with an element-wise tolerance of {"absoluteThreshold":0.000090957,"relativeThreshold":0}. 
PASS   SNR (85.58657894998035 dB) is greater than or equal to 85.58. 
PASS < [buffer-stitching-1] All assertions passed. (total 3 assertions) 
PASS > [buffer-stitching-2] Subsample buffer stitching, different rates 
PASS   Test 2: context.sampleRate is equal to 44100. 
FAIL X Stitched sine-wave buffers at sample rate 43800 does not equal [0,0.06264829635620117,0.12505052983760834,0.1869613528251648,0.24813774228096008,0.308339387178421,0.36732959747314453,0.4248766601085663,0.48075437545776367,0.53474360704422,0.5866317749023438,0.6362154483795166,0.6832994818687439,0.7276992797851562,0.769240140914917,0.8077590465545654...] with an element-wise tolerance of {"absoluteThreshold":0.0038986,"relativeThreshold":0}.
	Index	Actual			Expected		AbsError		RelError		Test threshold
	[30]	9.6648049354553223e-1	9.5236861705780029e-1	1.4111876487731934e-2	1.4817662231803117e-2	3.8985999999999999e-3
	[60]	-5.5004155635833740e-1	-5.8084774017333984e-1	3.0806183815002441e-2	5.3036590631839402e-2	3.8985999999999999e-3
	[90]	-6.1754375696182251e-1	-5.9811067581176758e-1	1.9433081150054932e-2	3.2490777937845654e-2	3.8985999999999999e-3
	[151]	2.0081890746951103e-2	-4.1306052356958389e-2	6.1387943103909492e-2	1.4861730812087184e+0	3.8985999999999999e-3
	[181]	-9.5454013347625732e-1	-9.3895959854125977e-1	1.5580534934997559e-2	1.6593402910202976e-2	3.8985999999999999e-3
	...and 1222 more errors.
	Max AbsError of 6.3063371824682690e-2 at index of 6615.
	[6615]	-6.3076853752136230e-2	-1.3481927453540266e-5	6.3063371824682690e-2	4.6776228430247684e+3	3.8985999999999999e-3
	Max RelError of 4.6776228430247684e+3 at index of 6615.
	[6615]	-6.3076853752136230e-2	-1.3481927453540266e-5	6.3063371824682690e-2	4.6776228430247684e+3	3.8985999999999999e-3
 assert_true: expected true got false
FAIL X SNR (43.58852204214277 dB) is not greater than or equal to 65.737. Got 43.58852204214277. assert_true: expected true got false
FAIL < [buffer-stitching-2] 2 out of 3 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 2 tasks were failed. assert_true: expected true got false
